researchOS/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9
httpx==0.27.0
anthropic==0.34.2
jinja2==3.1.4
itsdangerous==2.2.0
bcrypt==4.1.3
pyyaml==6.0.2
markdown==3.7
playwright>=1.44.0
requests>=2.31.0
crawl4ai>=0.8.0
